# service-a/target/classes/static/openapi.yaml
openapi: 3.0.3
info:
  title: demo-api
  version: 1.0.0
  description: Minimal API with two endpoints for logging/ELK testing.
servers:
  - url: http://localhost:8080
tags:
  - name: demo

paths:
  /api/hello:
    get:
      tags: [demo]
      summary: Greeting endpoint
      parameters:
        - name: name
          in: query
          description: Name to greet
          required: false
          schema:
            type: string
            default: world
      responses:
        "200":
          description: Greeting message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
              examples:
                example:
                  value:
                    message: "Hello, Labib!"
                    endpoint: "/api/hello"
                    ts: 1730625600000

  /api/compute:
    get:
      tags: [demo]
      summary: Small CPU task to vary latency and errors
      parameters:
        - name: n
          in: query
          description: Fibonacci steps (clamped 1–30)
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 30
      responses:
        "200":
          description: Computation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputeResponse'
              examples:
                example:
                  value:
                    result: 10946
                    endpoint: "/api/compute"
                    ts: 1730625601234
                    latency_ms: 182
        "500":
          description: Simulated failure (~5% of calls)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                example:
                  value:
                    error: simulated_failure
                    endpoint: /api/compute

components:
  schemas:
    HelloResponse:
      type: object
      properties:
        message: { type: string }
        endpoint: { type: string, example: /api/hello }
        ts: { type: integer, format: int64, description: epoch millis }
      required: [message, endpoint, ts]

    ComputeResponse:
      type: object
      properties:
        result: { type: integer }
        endpoint: { type: string, example: /api/compute }
        ts: { type: integer, format: int64, description: epoch millis }
        latency_ms: { type: integer }
      required: [result, endpoint, ts, latency_ms]

    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        endpoint: { type: string }
      required: [error, endpoint]